+++ CHANGELOG.md
@@ -2,16 +2,19 @@
 
 ## UNRELEASED
 
-* SCSensitivityLabel
-  * Added new parameters
-
 * EXOMobileDeviceMailboxPolicy
   * Fixed an error where if no MinPasswordLength was specified
     the Set-TargetResource threw n error trying to create a new
     policy;
+* EXOInboundConnector
+  * Fixed an issue where ResourceName was null during the export;
+* EXOOutboundConnector
+  * Fixed an issue where ResourceName was null during the export;
 * ODSettings
   * Fixed an issue where the GrooveBlockOption setting was never
     set properly;
+* SCSensitivityLabel
+  * Added new parameters
 
 ## 1.20.730.2